Add conda role
Conda should be installed by run.sh script, but it is currently failing.
This seems to be due to exceeding the 50-character maximum install path
(prefix). However, it isn't clear how to avoid the problem nor is it
clear the solution wouldn't cause other problems. Thus, the easiest
thing seems to be to install conda in advance.